requirements.txt
addict==2.4.0
appdirs==1.4.4
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
black==22.3.0
bleach==5.0.0
blinker==1.4
cached-property==1.5.2
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-completion==0.5.2
click-didyoumean==0.3.0
cloud-init==23.2.1
colorama==0.4.4
colorful==0.5.0
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docutils==0.18.1
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.7.0
flake8==4.0.1
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
icecream==2.1.2
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.1
mc-bin-client==1.0.1
mccabe==0.6.1
more-itertools==8.10.0
mypy==0.950
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openpyxl==3.1.2
packaging==21.3
pandas==2.2.1
pathspec==0.9.0
pexpect==4.8.0
pipenv==2022.5.2
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
psutil==5.9.1
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylru==1.2.1
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
pytest==7.1.2
pytest-click==1.1.0
python-apt==2.4.0+ubuntu2
python-box==4.2.3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
rulekit @ git+https://ghe.eng.fireeye.com/Helix-Detection/rulekit@129e35667ac56ce3e6d9bf70da7bc6de800b8477
schedule==1.1.0
SecretStorage==3.3.1
semver==3.0.0.dev3
service-identity==18.1.0
shell==1.0.1
shellingham==1.4.0
six==1.16.0
sos==4.4
ssh-import-id==5.11
stringcase==1.2.0
systemd-python==234
tabulate==0.8.9
tap==0.2
texttable==1.7.0
toml==0.10.2
tomli==2.0.1
twine==4.0.0
Twisted==22.1.0
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.4
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
wadllib==1.3.6
webencodings==0.5.1
xkit==0.0.0
zipp==3.8.0
zope.interface==5.4.0
